create function for music player with slider

+++ lib/feature/presentation/blocs/detail_music/detail_music_state.dart
@@ -2,7 +2,7 @@ part of 'detail_music_bloc.dart';
 
 abstract class DetailMusicState extends Equatable {
   final SearchResultModel? resultModel;
-  final double? musicIndex;
+  final int? musicIndex;
   final bool? isPlayed;
   const DetailMusicState({required this.resultModel, required this.musicIndex, required this.isPlayed});
 
@@ -16,6 +16,6 @@ class DetailMusicInitial extends DetailMusicState {
 }
 
 class DetailMusicLoaded extends DetailMusicState {
-  const DetailMusicLoaded({SearchResultModel? resultModel, double? musicIndex, bool? isPlayed})
+  const DetailMusicLoaded({SearchResultModel? resultModel, int? musicIndex, bool? isPlayed})
       : super(resultModel: resultModel, musicIndex: musicIndex, isPlayed: isPlayed);
 }
